feat: extract map bounds/walkability into core/engine/game_map_ops.js and wire in core/game.js
