+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/game.js
@@ -72,6 +72,7 @@ import { log as logFacade } from "./facades/log.js";
 import { int as rngInt, chance as rngChance, float as rngFloat } from "./facades/rng.js";
 import { createGameCombatOps } from "./engine/game_combat_ops.js";
 import { createGameInventoryOps } from "./engine/game_inventory_ops.js";
+import { createGameMapOps } from "./engine/game_map_ops.js";
 import { setupInputBridge, initUIHandlersBridge } from "./engine/game_ui_bridge.js";
 // Side-effect import to ensure FollowersItems attaches itself to window.FollowersItems
 import "./followers_items.js";
@@ -358,7 +359,10 @@ import "./sandbox/runtime.js";
     return null;
   }
 
+  const mapOps = createGameMapOps(getCtx);
+
   // RNG helpers via facade
+
   const randInt = (min, max) => rngInt(min, max, rng);
   const chance = (p) => rngChance(p, rng);
   const capitalize = ((typeof window !== "undefined" && window.PlayerUtils && typeof window.PlayerUtils.capitalize === "function")
@@ -416,64 +420,16 @@ import "./sandbox/runtime.js";
   }
 
   function inBounds(x, y) {
-    // Centralize via Utils.inBounds; fallback to local map bounds
-    const U = modHandle("Utils");
-    if (U && typeof U.inBounds === "function") {
-      return !!U.inBounds(getCtx(), x, y);
-    }
-    const rows = Array.isArray(map) ? map.length : 0;
-    const cols = rows && Array.isArray(map[0]) ? map[0].length : 0;
-    return x >= 0 && y >= 0 && x < cols && y < rows;
+    return mapOps.inBounds(x, y);
   }
 
   
   
 
-  
+
 
   function isWalkable(x, y) {
-    // Upstairs overlay-aware walkability: when active and inside the inn interior, honor upstairs tiles.
-    try {
-      if (innUpstairsActive && tavern && innUpstairs) {
-        const b = tavern.building || null;
-        const up = innUpstairs;
-        if (b && up) {
-          const ox = up.offset ? up.offset.x : (b.x + 1);
-          const oy = up.offset ? up.offset.y : (b.y + 1);
-          const lx = x - ox, ly = y - oy;
-          const w = up.w | 0, h = up.h | 0;
-          if (lx >= 0 && ly >= 0 && lx < w && ly < h) {
-            const row = up.tiles && up.tiles[ly];
-            const t = row ? row[lx] : null;
-            if (t != null) {
-              // Treat WALL as not walkable; allow FLOOR and STAIRS; disallow DOOR upstairs to avoid "walkable doors" issue.
-              return t === TILES.FLOOR || t === TILES.STAIRS;
-            }
-          }
-        }
-      }
-    } catch (_) {}
-
-    // Centralize via Utils.isWalkableTile; fallback to tile-type check
-    const U = modHandle("Utils");
-    if (U && typeof U.isWalkableTile === "function") {
-      return !!U.isWalkableTile(getCtx(), x, y);
-    }
-    const rows = Array.isArray(map) ? map.length : 0;
-    const cols = rows && Array.isArray(map[0]) ? map[0].length : 0;
-    if (x < 0 || y < 0 || x >= cols || y >= rows) return false;
-    const t = map[y][x];
-    // Fallback walkability when Utils.isWalkableTile is unavailable.
-    // Keep in sync with utils.isWalkableTile for town/dungeon maps.
-    return (
-      t === TILES.FLOOR ||
-      t === TILES.DOOR ||
-      t === TILES.STAIRS ||
-      t === TILES.ROAD ||
-      t === TILES.PIER ||
-      t === TILES.SHIP_DECK ||
-      t === TILES.SHIP_EDGE
-    );
+    return mapOps.isWalkable(x, y);
   }
