+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/docs/index.html
@@ -125,6 +125,7 @@
         { title: 'Engineering: UIOrchestration Modularization (Slice A)', path: 'docs/ui_orchestration_modularization.md' },
         { title: 'Engineering: UIBridge Modularization (Slice B)', path: 'docs/ui_bridge_modularization.md' },
         { title: 'Engineering: Boot / Load-Order Modularization (Slice C)', path: 'docs/boot_load_order_modularization.md' },
+        { title: 'Engineering: core/game.js Shrink — Map Ops (Slice D)', path: 'docs/game_js_shrink_slice_map_ops.md' },
 
         { title: 'Data Docs: Index', path: 'data/docs/README.md' },
         { title: 'Data Docs: Palette Schema', path: 'data/docs/palette_schema.md' },
@@ -167,6 +168,7 @@
         'docs/ui_orchestration_modularization.md': 'UIOrchestration file split: new module layout + invariants.',
         'docs/ui_bridge_modularization.md': 'UIBridge file split: new module layout + invariants.',
         'docs/boot_load_order_modularization.md': 'Boot sequence refactor: src/main.js now imports grouped boot manifests.',
+        'docs/game_js_shrink_slice_map_ops.md': 'core/game.js shrink: extract inBounds/isWalkable policy into core/engine/game_map_ops.js.',
 
         'data/docs/README.md': 'Index of data documentation and palette guides.',
         'data/docs/palette_schema.md': 'Schema for palette JSON: keys, sections, and constraints.',
